packetpool: allow larger max-pending-packets
Original limit was due to a specific data structure.

+++ src/source-pcap-file.c
@@ -33,7 +33,7 @@
 #include "suricata.h"
 #include "conf.h"
 
-extern uint16_t max_pending_packets;
+extern uint32_t max_pending_packets;
 PcapFileGlobalVars pcap_g;
 
 /**
